Modified targets vector from tfModules to character
Modified the "targets" that are used from correlation matrix to check if TF and targets are correlated to be transformed into character vector before using them to subset the corrMat. Ran into issues where instead of using the gene names to subset, R was using integers and used values from other genes to assess correlation. This led to almost random assignment of 1, 0 or -1 to the target genes which directly affects the downstream enrichment calculations!

This might go unnoticed if (stringsAsFactors = FALSE) is set as a global option but would create wrong output for users that do not have this setting enabled!